+++ CHANGELOG.md
@@ -14,6 +14,7 @@ v2.0.0-rc.16
  * ons-fab: Fix [#1496](https://github.com/OnsenUI/OnsenUI/issues/1496).
  * ons-list-item: Fix [#1499](https://github.com/OnsenUI/OnsenUI/issues/1499)
  * ons-tabbar: Fix [#1501](https://github.com/OnsenUI/OnsenUI/issues/1501)
+ * ons-carousel: initial-index works when the carousel is inside ons-navigator.
 
 v2.0.0-rc.15
 ----